requirements.txt
tensorflow >= 1.12,<1.13.0
pillow >= 5.0
flask >= 1.0
